+++ PHYS495_CircuitAnalyzer/src/circuit_logic/Inductor.java
@@ -1,12 +1,12 @@
 package circuit_logic;
 
 public class Inductor extends CircuitComponent{
-	public Inductor(float value) {
+	public Inductor(double value) {
 		super("Inductor",value);
 	}
 
 	public ComplexNumber calculateImpedance(double frequency) {
-		return new ComplexNumber((float)0,(float)frequency * value);
+		return new ComplexNumber(0,frequency * value);
 	}
 	
 }
