+++ PHYS495_CircuitAnalyzer/src/circuit_logic/JunctionStart.java
@@ -2,13 +2,13 @@ package circuit_logic;
 
 public class JunctionStart extends CircuitComponent{
 	
-	public JunctionStart (float value) {
+	public JunctionStart (double value) {
 		super("JunctionStart",value);
 	}
 
 
 	public ComplexNumber calculateImpedance(double frequency) {
-		return new ComplexNumber((float)0,(float)0);
+		return new ComplexNumber(0,0);
 	}
 
 }
