+++ PHYS495_CircuitAnalyzer/src/circuit_logic/Resistor.java
@@ -1,14 +1,14 @@
 package circuit_logic;
 
 public class Resistor extends CircuitComponent{
-	public Resistor(float value) {
+	public Resistor(double value) {
 		super("Resistor",value);
 	}
 	
 	//units:Ohm (O),kiloOhms (K),MegaOhms(M)
 
 	public ComplexNumber calculateImpedance(double frequency) {
-		return new ComplexNumber ((float)value,(float)0);
+		return new ComplexNumber (value,0);
 	}
 	
 }
